ui-redesign: Dashboard
